feat: refactor rule panel popover w editors/helpers/constants + hook to support rule/condition editing
